requirements.txt
langchain==0.1.12
langchain-community==0.0.28
langchain-openai==0.0.8
faiss-cpu
python-dotenv==1.0.1
PyMuPDF==1.23.26
streamlit==1.32.2
# 强行锁定版本，避开最新版的 Bug
transformers==4.40.0
sentence-transformers==2.7.0
pypdf
